refactor(frontend): extract the document open flow into a hook
The chat detail and user detail screens both need to open an original
document, and the popup-blocker workaround is too easy to get wrong to
copy into three places.

- Move the open flow out of Documents into useOpenDocument, keeping the
  blank tab that has to be opened before the await and the opener reset.
- Take the failure message through a callback so each screen keeps
  reporting errors in its own banner instead of the hook owning one.
